New uniform shutdown/reboot and queue cycle
The queue now has a section marker "##" marking those points in it after
which a dijunct section starts, so that cqappend.bsh can skip over
intermediate sections where possible. Hydra now passes a "HYDRA" key so that
programs know whether they run alone or not and can thus decide on whether to
put a section marker or not.
The wakeup was migrated into a separate script which is now called from
within the queue (though as a local command) like all other commands, thus
allowing for a uniform wakeup and shutdown cycle without inconvenient
sideeffects (such as a computer shutting down immediately after it was booted
because the wakeup was emitted out of line).